refactor: 새로 구현한 CouponFixture로 테스트 픽스처 대체
