+++ backend/src/test/java/com/woowacourse/kkogkkog/coupon/domain/repository/CouponRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.coupon.domain.repository;
 
-import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFixture.COFFEE;
+import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFixture.createCoupon;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.RECEIVER;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.RECEIVER2;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.SENDER;
@@ -66,9 +66,9 @@ class CouponRepositoryTest {
         @DisplayName("보낸 사람의 기준으로 쿠폰 목록을 조회할 수 있다.")
         @Test
         void senderSuccess() {
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver2));
-            couponRepository.save(COFFEE.getCoupon(receiver, sender));
+            couponRepository.save(createCoupon(sender, receiver));
+            couponRepository.save(createCoupon(sender, receiver2));
+            couponRepository.save(createCoupon(receiver, sender));
             entityManager.flush();
             entityManager.clear();
 
@@ -79,9 +79,9 @@ class CouponRepositoryTest {
         @DisplayName("보낸 사람과 상태를 기준으로 쿠폰 목록을 조회할 수 있다.")
         @Test
         void success_senderWithStatus() {
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            Coupon coupon = COFFEE.getCoupon(sender, receiver);
+            couponRepository.save(createCoupon(sender, receiver));
+            couponRepository.save(createCoupon(sender, receiver));
+            Coupon coupon = createCoupon(sender, receiver);
             couponRepository.save(coupon);
             coupon.changeState(
                 new CouponEvent(CouponEventType.REQUEST, LocalDateTime.now().plusDays(1L)),
@@ -97,9 +97,9 @@ class CouponRepositoryTest {
         @DisplayName("받은 사람의 기준으로 쿠폰 목록을 조회할 수 있다.")
         @Test
         void receiverSuccess() {
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver2));
-            couponRepository.save(COFFEE.getCoupon(receiver, sender));
+            couponRepository.save(createCoupon(sender, receiver));
+            couponRepository.save(createCoupon(sender, receiver2));
+            couponRepository.save(createCoupon(receiver, sender));
             couponRepository.flush();
 
             Slice<Coupon> actual = couponRepository.findAllByReceiver(receiver, PageRequest.of(0, 5));
@@ -109,9 +109,9 @@ class CouponRepositoryTest {
         @DisplayName("받은 사람과 상태를 기준으로 쿠폰 목록을 조회할 수 있다.")
         @Test
         void success_receiverWithStatus() {
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver));
-            Coupon coupon = COFFEE.getCoupon(sender, receiver);
+            couponRepository.save(createCoupon(sender, receiver));
+            couponRepository.save(createCoupon(sender, receiver));
+            Coupon coupon = createCoupon(sender, receiver);
             couponRepository.save(coupon);
             coupon.changeState(
                 new CouponEvent(CouponEventType.REQUEST, LocalDateTime.now().plusDays(1L)),
@@ -145,14 +145,14 @@ class CouponRepositoryTest {
         @Test
         void acceptedCouponsAfterCurrentTime() {
             // given
-            couponRepository.save(COFFEE.getCoupon(sender, receiver,
+            couponRepository.save(createCoupon(sender, receiver,
                 new CouponState(CouponStatus.ACCEPTED, LocalDateTime.now().plusDays(1))));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver2,
+            couponRepository.save(createCoupon(sender, receiver2,
                 new CouponState(CouponStatus.ACCEPTED, LocalDateTime.now())));
 
-            couponRepository.save(COFFEE.getCoupon(sender, receiver,
+            couponRepository.save(createCoupon(sender, receiver,
                 new CouponState(CouponStatus.REQUESTED, LocalDateTime.now().plusDays(2))));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver,
+            couponRepository.save(createCoupon(sender, receiver,
                 new CouponState(CouponStatus.FINISHED, LocalDateTime.now().plusDays(2))));
 
             // when
@@ -168,9 +168,9 @@ class CouponRepositoryTest {
         @Test
         void couponsBeforeTheCurrentTime() {
             // given
-            couponRepository.save(COFFEE.getCoupon(sender, receiver,
+            couponRepository.save(createCoupon(sender, receiver,
                 new CouponState(CouponStatus.ACCEPTED, LocalDateTime.now().minusDays(1))));
-            couponRepository.save(COFFEE.getCoupon(sender, receiver,
+            couponRepository.save(createCoupon(sender, receiver,
                 new CouponState(CouponStatus.ACCEPTED, LocalDateTime.now().minusDays(2))));
 
             // when
