+++ backend/src/test/java/com/woowacourse/kkogkkog/infrastructure/application/WoowacoursePushAlarmListenerTest.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.infrastructure.application;
 
-import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFixture.COFFEE;
+import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFixture.createCoupon;
 import static com.woowacourse.kkogkkog.support.fixture.dto.CouponDtoFixture.COFFEE_쿠폰_저장_요청;
 import static com.woowacourse.kkogkkog.support.fixture.dto.CouponDtoFixture.쿠폰_상태_변경_요청;
 import static org.mockito.BDDMockito.anyString;
@@ -82,7 +82,7 @@ public class WoowacoursePushAlarmListenerTest {
         @DisplayName("쿠폰 상태를 변경할 때, woowacourse 워크스페이스로 슬랙 푸시 알림을 보낸다.")
         void success_couponUpdate() {
             //given
-            Coupon coupon = couponRepository.save(COFFEE.getCoupon(sender, receiver));
+            Coupon coupon = couponRepository.save(createCoupon(sender, receiver));
             given(woowacourseUserRepository.contains(anyString())).willReturn(true);
             given(woowacourseUserRepository.getUserId(anyString())).willReturn(Optional.of("userId"));
